+font

+++ red-triplane-fonts-api/src/com/jfixby/r3/ext/api/font/CharRasterInfo.java
@@ -3,9 +3,9 @@ package com.jfixby.r3.ext.api.font;
 public class CharRasterInfo {
 
 	public String string_value;
-	public int tile_width_in_pixels;
-	public int tile_height_in_pixels;
-	public int font_size;
+	public float tile_width_in_pixels;
+	public float tile_height_in_pixels;
+	public float font_size;
 	public float tile_origin_x;
 	public float tile_origin_y;
 	public String raster_id;
